Fix colorbox not updating on scheme apply + fix input bug
The colorpickers' values are now explicitly set after the scheme is
applied, therefore matching the values of the scheme. The input field
now has a mutable value, thus rendering the event listeners on the
colorpicker redundant. So changing the input will automatically update
the corresponding color property in the configuration service.
Furthermore, this fixes the input field being buggy in that sense that
the function called in the event of a color change was setting the
color property to undefined in some cases.